server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
requests==2.31.0
groq==0.9.0
gunicorn==22.0.0
